backend/requirements.txt
# Core Flask Framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-Cors==5.0.0

# Serialization & Validation
marshmallow==3.21.3
marshmallow-sqlalchemy==1.4.2

# Environment Management
python-dotenv==1.0.1

# Database Driver (for PostgreSQL; optional if using SQLite)
psycopg2-binary==2.9.9

# Testing & Debugging (optional but recommended)
pytest==8.3.3
pytest-flask==1.3.0
ipdb==0.13.13
